restructure acyclic funcs
