+++ src/core/structsos/structsos.py
@@ -1,11 +1,13 @@
 from typing import Union, Dict, Optional, Any
+from functools import partial
 
 import sympy as sp
 
 from .solution import SolutionStructural, SolutionStructuralSimple
 
-from .utils import Coeff
-from .sparse  import sos_struct_sparse, sos_struct_heuristic
+from .utils import Coeff, PolynomialNonpositiveError, PolynomialUnsolvableError
+from .sparse  import sos_struct_sparse, sos_struct_heuristic, sos_struct_extract_factors
+from .quadratic import sos_struct_quadratic, sos_struct_acyclic_quadratic
 from .cubic   import sos_struct_cubic
 from .quartic import sos_struct_quartic
 from .quintic import sos_struct_quintic
@@ -13,12 +15,13 @@ from .sextic  import sos_struct_sextic
 from .septic  import sos_struct_septic
 from .octic   import sos_struct_octic
 from .nonic   import sos_struct_nonic
-from .acyclic import sos_struct_acyclic
+from .acyclic import sos_struct_acyclic_sparse
 
 from ...utils import verify_hom_cyclic
 
 
 SOLVERS = {
+    2: sos_struct_quadratic,
     3: sos_struct_cubic,
     4: sos_struct_quartic,
     5: sos_struct_quintic,
@@ -28,9 +31,14 @@ SOLVERS = {
     9: sos_struct_nonic,
 }
 
+SOLVERS_ACYCLIC = {
+    2: sos_struct_acyclic_quadratic,
+}
+
 def _structural_sos_handler(
         coeff: Union[sp.polys.Poly, Coeff, Dict],
-        real = True,
+        real: bool = True,
+        is_cyc: bool = True
     ) -> sp.Expr:
     """
     Perform structural sos and returns an sympy expression. This function could be called 
@@ -41,18 +49,37 @@ def _structural_sos_handler(
     if not isinstance(coeff, Coeff):
         coeff = Coeff(coeff)
 
-    degree = coeff.degree()
 
-    # first try sparse cases
-    solution = sos_struct_sparse(coeff, recurrsion = _structural_sos_handler, real = real)
+    partial_recur = partial(_structural_sos_handler, is_cyc = is_cyc)
 
-    if solution is None:
-        solver = SOLVERS.get(degree, None)
-        if solver is not None:
-            solution = SOLVERS[degree](coeff, recurrsion = _structural_sos_handler, real = real)
+    if is_cyc:
+        prior_solver = sos_struct_sparse
+        solvers = SOLVERS
+        heuristic_solver = sos_struct_heuristic
+    else:
+        prior_solver = sos_struct_acyclic_sparse
+        solvers = SOLVERS_ACYCLIC
+        heuristic_solver = lambda *args, **kwargs: None
 
-    if solution is None and degree > 6:
-        solution = sos_struct_heuristic(coeff, recurrsion = _structural_sos_handler)
+    try:
+        solution = sos_struct_extract_factors(coeff, recurrsion = partial_recur, real = real)
+
+        degree = coeff.degree()
+
+        # first try sparse cases
+        if solution is None:
+            solution = prior_solver(coeff, recurrsion = partial_recur, real = real)
+
+        if solution is None:
+            solver = solvers.get(degree, None)
+            if solver is not None:
+                solution = solver(coeff, recurrsion = partial_recur, real = real)
+
+        if solution is None and degree > 6:
+            solution = heuristic_solver(coeff, recurrsion = partial_recur)
+
+    except PolynomialUnsolvableError:
+        return None
 
     return solution
 
@@ -85,10 +112,10 @@ def StructuralSOS(
     if not is_hom:
         return None
 
-    if is_cyc:
-        solution = _structural_sos_handler(Coeff(poly), real = real)
-    else:
-        solution = sos_struct_acyclic(Coeff(poly), real = real)
+    try:
+        solution = _structural_sos_handler(Coeff(poly), real = real, is_cyc = is_cyc)
+    except (PolynomialNonpositiveError, PolynomialUnsolvableError):
+        return None
 
     if solution is None:
         return None
