+++ src/core/structsos/utils.py
@@ -165,6 +165,14 @@ class Coeff():
         return d, new_coeff
 
 
+class StructuralSOSError(Exception): ...
+
+class PolynomialUnsolvableError(StructuralSOSError): ...
+
+class PolynomialNonpositiveError(PolynomialUnsolvableError): ...
+
+
+
 def radsimp(expr: Union[sp.Expr, List[sp.Expr]]) -> sp.Expr:
     """
     Rationalize the denominator by removing square roots. Wrapper of sympy.radsimp.
